+++ framework_src/src/org/smarty/core/support/jdbc/holder/MSSQLHolder.java
@@ -24,7 +24,7 @@ public class MSSQLHolder extends SQLHolder {
         StringBuilder sb = new StringBuilder(sql);
 
         // 计算总页数
-        Long pageCount = 0L;
+        int pageCount = 0;
         if (pager.getTotalCount() % pager.getPageSize() == 0) {
             pageCount = pager.getTotalCount() / pager.getPageSize();
         } else if (pager.getTotalCount() % pager.getPageSize() > 0) {
