+++ framework_src/src/org/smarty/core/support/jdbc/holder/OracleHolder.java
@@ -23,7 +23,7 @@ public class OracleHolder extends SQLHolder {
         sb.append(sql);
 
         // 计算总页数
-        Long pageCount = 0L;
+        int pageCount = 0;
         if (pager.getTotalCount() % pager.getPageSize() == 0) {
             pageCount = pager.getTotalCount() / pager.getPageSize();
         } else if (pager.getTotalCount() % pager.getPageSize() > 0) {
